@@ -25,53 +25,52 @@
     "node": ">=18.12.0"
   },
   "dependencies": {
-    "@angular/animations": "^20.0.3",
+    "@angular/animations": "^20.0.4",
     "@angular/cdk": "^20.0.3",
-    "@angular/common": "^20.0.3",
-    "@angular/compiler": "^20.0.3",
-    "@angular/core": "^20.0.3",
-    "@angular/forms": "^20.0.3",
+    "@angular/common": "^20.0.4",
+    "@angular/compiler": "^20.0.4",
+    "@angular/core": "^20.0.4",
+    "@angular/forms": "^20.0.4",
     "@angular/material": "^20.0.3",
-    "@angular/platform-browser": "^20.0.3",
-    "@angular/platform-browser-dynamic": "^20.0.3",
-    "@angular/platform-server": "^20.0.3",
-    "@angular/router": "^20.0.3",
-    "@angular/ssr": "^20.0.2",
+    "@angular/platform-browser": "^20.0.4",
+    "@angular/platform-browser-dynamic": "^20.0.4",
+    "@angular/platform-server": "^20.0.4",
+    "@angular/router": "^20.0.4",
+    "@angular/ssr": "^20.0.3",
     "@stackblitz/sdk": "^1.11.0",
-    "domhandler": "^5.0.3",
     "express": "^4.21.2",
     "lodash-es": "^4.17.21",
     "maplibre-gl": "^5.0.0",
-    "rxjs": "7.8.1",
+    "rxjs": "^7.8.2",
     "scroll-into-view-if-needed": "^3.1.0",
     "tslib": "^2.6.3"
   },
   "devDependencies": {
-    "@angular-eslint/builder": "^18.3.0",
-    "@angular-eslint/eslint-plugin": "^18.3.0",
-    "@angular-eslint/eslint-plugin-template": "^18.3.0",
-    "@angular-eslint/schematics": "^18.3.0",
-    "@angular-eslint/template-parser": "^18.3.0",
-    "@angular/build": "^20.0.2",
-    "@angular/cli": "^20.0.2",
-    "@angular/compiler-cli": "^20.0.3",
+    "@angular-eslint/builder": "^20.1.1",
+    "@angular-eslint/eslint-plugin": "^20.1.1",
+    "@angular-eslint/eslint-plugin-template": "^20.1.1",
+    "@angular-eslint/schematics": "^20.1.1",
+    "@angular-eslint/template-parser": "^20.1.1",
+    "@angular/build": "^20.0.3",
+    "@angular/cli": "^20.0.3",
+    "@angular/compiler-cli": "^20.0.4",
     "@cypress/schematic": "^2.5.2",
-    "@shellygo/cypress-test-utils": "^2.1.30",
-    "@types/express": "^4.17.21",
+    "@shellygo/cypress-test-utils": "^5.0.2",
+    "@types/express": "^5.0.3",
     "@types/geojson": "^7946.0.14",
     "@types/jasmine": "~5.1.4",
     "@types/jasminewd2": "^2.0.13",
     "@types/lodash-es": "^4.17.12",
-    "@types/node": "^22.5.0",
+    "@types/node": "^24.0.4",
     "@types/pixelmatch": "^5.2.6",
     "@types/pngjs": "^6.0.5",
-    "@typescript-eslint/eslint-plugin": "^8.2.0",
-    "@typescript-eslint/parser": "^8.2.0",
+    "@typescript-eslint/eslint-plugin": "^8.33.1",
+    "@typescript-eslint/parser": "^8.33.1",
     "angular-cli-ghpages": "^2.0.1",
     "auto-changelog": "^2.5.0",
     "commitizen": "^4.3.0",
-    "cypress": "^14.4.1",
-    "eslint": "^8.57.0",
+    "cypress": "^14.5.0",
+    "eslint": "^9.28.0",
     "jasmine-core": "~5.2.0",
     "jasmine-spec-reporter": "~7.0.0",
     "karma": "~6.4.4",
@@ -81,8 +80,8 @@
     "karma-jasmine-html-reporter": "^2.1.0",
     "karma-mocha-reporter": "^2.2.5",
     "lint-staged": "^15.2.9",
-    "ng-packagr": "^20.0.0",
-    "pixelmatch": "^6.0.0",
+    "ng-packagr": "^20.0.1",
+    "pixelmatch": "^7.1.0",
     "pngjs": "^7.0.0",
     "standard-version": "^9.5.0",
     "ts-node": "^10.9.2",
